js clean
