+++ app/javascript/components/button_back.js
@@ -1,9 +1,12 @@
 import $ from "jquery";
 
 const buttonBack = () => {
-  $("btn-3d").click(function(event){
-    event.preventDefault();
+  const buttonBack = document.querySelector(".btn-3d")
+  if (buttonBack) {
+    $(".btn-3d").click(function(event){
+      event.preventDefault();
     });
+  }
 }
 
 export { buttonBack }
