+++ app/javascript/packs/application.js
@@ -1,26 +1,14 @@
 import "bootstrap";
-
+import initPolar from "../components/dashboard";
 import { toggleNav } from "../components/vertical_navbar"
-toggleNav();
-
-
-import { initStarRating } from "../plugins/init_star_rating";
-initStarRating();
-
+import { initStarRating } from "../components/init_star_rating";
 import { buttonBack } from "../components/button_back";
-buttonBack();
-
 import { autocompleteSearch } from "../components/autocomplete"
-autocompleteSearch();
-
 import { animateChallengeButton } from "../components/button_challenge";
+
+initPolar();
+toggleNav();
+initStarRating();
+buttonBack();
+autocompleteSearch();
 animateChallengeButton();
-
-// import initPolar from "../components/dashboard";
-// initPolar();
-//import { deployDescription } from "../components/card_category";
-//deployDescription();
-
-//import { initVerticalNavbar } from "../components/vertical_navbar"
-//initVerticalNavbar();
-
