--- app/javascript/plugins/init_star_rating.js
@@ -1,14 +1,0 @@
-import "jquery-bar-rating";
-import "jquery-bar-rating/dist/themes/css-stars.css";
-
-const initStarRating = () => {
-  $('.modal').each(function() {
-    $( this ).on('shown.bs.modal', function(){
-      $('.modal-rating').barrating({
-        theme: 'css-stars'
-      });
-    });
-  });
-};
-
-export { initStarRating };
